api/requirements.txt
pyodbc
flask
python-dotenv